refactor: remove App, main, router tsx files
